# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BigNumber": {},
...
			"MastersTraitType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layers": {
						"items": {
							"$ref": "#/components/schemas/TraitLayer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TraitLayer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layer": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MastersTraitType"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TraitGender": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersColor": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"typeId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"color",
					"typeId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersTrait": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.926Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.926Z"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"frontImage": {
						"type": "string"
					},
					"backImage": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MastersTraitType"
					},
					"gender": {
						"$ref": "#/components/schemas/TraitGender"
					},
					"color": {
						"$ref": "#/components/schemas/MastersColor"
					},
					"hide": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"type",
					"gender",
					"color",
					"hide"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ItemType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"layers": {
						"items": {
							"$ref": "#/components/schemas/ItemLayer"
						},
						"type": "array"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/ItemCategory"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"layers",
					"categories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemLayer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layer": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ItemType"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layer",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ItemType"
					}
				},
				"required": [
					"id",
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.937Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.937Z"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/ItemCategory"
					},
					"gender": {
						"$ref": "#/components/schemas/TraitGender"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"frontImage": {
						"type": "string"
					},
					"backImage": {
						"type": "string"
					},
					"maskImage": {
						"type": "string"
					},
					"rarity": {
						"type": "string"
					},
					"effects": {
						"type": "string"
					},
					"storyNotes": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"category",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MastersItemBalance": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.946Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.947Z"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/ItemCategory"
					},
					"gender": {
						"$ref": "#/components/schemas/TraitGender"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"frontImage": {
						"type": "string"
					},
					"backImage": {
						"type": "string"
					},
					"maskImage": {
						"type": "string"
					},
					"rarity": {
						"type": "string"
					},
					"effects": {
						"type": "string"
					},
					"storyNotes": {
						"type": "string"
					},
					"balance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"category",
					"image",
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersAvatar": {
				"properties": {
					"tokenId": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"attributes": {
						"items": {
							"properties": {
								"value": {
									"type": "string"
								},
								"trait_type": {
									"type": "string"
								}
							},
							"required": [
								"value",
								"trait_type"
							],
							"type": "object"
						},
						"type": "array"
					},
					"equippedItems": {
						"items": {
							"$ref": "#/components/schemas/MastersItem"
						},
						"type": "array"
					},
					"traits": {
						"items": {
							"$ref": "#/components/schemas/MastersTrait"
						},
						"type": "array"
					},
					"layers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"tokenId",
					"image",
					"name",
					"attributes",
					"equippedItems",
					"traits",
					"layers"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "masters-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/masters/items/mint": {
			"post": {
				"operationId": "MintMastersItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									},
									"receiver": {
										"type": "string"
									},
									"itemId": {
										"type": "number",
										"format": "double"
									},
									"clientId": {
										"type": "string"
									},
									"apiKey": {
										"type": "string"
									}
								},
								"required": [
									"amount",
									"receiver",
									"itemId",
									"clientId",
									"apiKey"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/masters/items/categories": {
			"get": {
				"operationId": "GetMastersItemCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ItemCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/items/balance/{address}": {
			"get": {
				"operationId": "GetMastersItemsBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersItemBalance"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "includeSubnet",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/masters/items/{tokenId}": {
			"get": {
				"operationId": "GetMastersItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MastersItem"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/items": {
			"get": {
				"operationId": "GetMastersItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars/image/{tokenId}": {
			"post": {
				"operationId": "UpdateImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "extension",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars/image-upload-url/{tokenId}": {
			"get": {
				"operationId": "GetMastersPFPImageUploadURL",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"tokenId": {
											"type": "number",
											"format": "double"
										},
										"layers": {
											"items": {},
											"type": "array"
										},
										"uploadUrl": {
											"type": "string"
										}
									},
									"required": [
										"tokenId",
										"layers",
										"uploadUrl"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "extension",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/masters/avatars/images/{address}": {
			"get": {
				"operationId": "GetMastersAvatarImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"profilePictures": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"name": {
											"type": "string"
										},
										"type": {
											"type": "string"
										}
									},
									"required": [
										"profilePictures",
										"name",
										"type"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars/balance/{address}": {
			"get": {
				"operationId": "GetMastersAvatarBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars/prices": {
			"get": {
				"operationId": "GetMastersAvatarPrices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars/free": {
			"post": {
				"operationId": "GetMastersFreePFPSignature",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"signature": {
											"type": "string"
										},
										"nonce": {
											"$ref": "#/components/schemas/BigNumber"
										},
										"traits": {
											"items": {
												"type": "number",
												"format": "double"
											},
											"type": "array"
										},
										"receiver": {
											"type": "string"
										}
									},
									"required": [
										"signature",
										"nonce",
										"traits",
										"receiver"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"traits": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									},
									"receiver": {
										"type": "string"
									}
								},
								"required": [
									"traits"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/masters/avatars/{tokenId}": {
			"get": {
				"operationId": "GetMastersAvatar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MastersAvatar"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/avatars": {
			"post": {
				"operationId": "GetMastersPFPSignature",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"payWithTicket": {
										"type": "boolean"
									},
									"currency": {
										"type": "string"
									},
									"traits": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									},
									"receiver": {
										"type": "string"
									}
								},
								"required": [
									"payWithTicket",
									"currency",
									"traits"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}